% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHS Payslip Function.R
\name{nhs_payslip}
\alias{nhs_payslip}
\title{Calculate NHS Take-Home Pay}
\usage{
nhs_payslip(
  annual_wage,
  income_tax_allowance = 12570,
  student_loan_plan = "none",
  period = "monthly"
)
}
\arguments{
\item{annual_wage}{The annual wage before tax and other deductions.}

\item{income_tax_allowance}{The income tax allowance to be used. Defaults to the UK standard of 12570.}

\item{student_loan_plan}{Options are 1,2,4,5,'postgraduate' or 'none'. Defaults to 'none'.}

\item{period}{Specify whether the output should be the annual breakown, or a monthly breakdown. Accepts 'monthly' or 'annual'. Defaults to 'monthly'. Monthly breakdown is equivalent to 1/12th of annual values.}
}
\description{
This function provides the value for take home pay, income tax, national insurance, pension contributions and  student loan repayments for staff on NHS AFC pay conditions.
}
\examples{
nhs_payslip(annual_wage = 50000)
}
